NPE check in case if file path not initialised

+++ flow-tests/test-live-reload/src/test/java/com/vaadin/flow/uitest/ui/ThemeLiveReloadIT.java
@@ -146,9 +146,11 @@ public class ThemeLiveReloadIT extends AbstractLiveReloadIT {
     }
 
     private void deleteFile(String filePath) {
-        File fileToDelete = new File(filePath);
-        if (fileToDelete.exists() && !fileToDelete.delete()) {
-            Assert.fail("Unable to delete " + filePath);
+        if (filePath != null) {
+            File fileToDelete = new File(filePath);
+            if (fileToDelete.exists() && !fileToDelete.delete()) {
+                Assert.fail("Unable to delete " + filePath);
+            }
         }
     }
 }
